Split the rest of the PLP into PLPContents.
